tftypes: Change (AttributePath).WithElementKeyInt() method parameter to int (#101)
* tftypes: Add (AttributePath).WithElementKeyInt64() method, change WithElementKeyInt() method parameter to int

Reference: https://github.com/hashicorp/terraform-plugin-go/issues/59

* Update CHANGELOG for #101

* tftypes: Remove (AttributePath).WithElementKeyInt64()

* Update CHANGELOG for #101

+++ tftypes/list.go
@@ -71,9 +71,9 @@ func valueFromList(typ Type, in interface{}) (Value, error) {
 		var valType Type
 		for pos, v := range value {
 			if v.Type().Is(DynamicPseudoType) && v.IsKnown() {
-				return Value{}, NewAttributePath().WithElementKeyInt(int64(pos)).NewErrorf("invalid value %s for %s", v, v.Type())
+				return Value{}, NewAttributePath().WithElementKeyInt(pos).NewErrorf("invalid value %s for %s", v, v.Type())
 			} else if !v.Type().Is(DynamicPseudoType) && !v.Type().UsableAs(typ) {
-				return Value{}, NewAttributePath().WithElementKeyInt(int64(pos)).NewErrorf("can't use %s as %s", v.Type(), typ)
+				return Value{}, NewAttributePath().WithElementKeyInt(pos).NewErrorf("can't use %s as %s", v.Type(), typ)
 			}
 			if valType == nil {
 				valType = v.Type()
